+++ hutool-core/src/main/java/cn/hutool/core/convert/AbstractConverter.java
@@ -1,6 +1,7 @@
 package cn.hutool.core.convert;
 
 import cn.hutool.core.util.ArrayUtil;
+import cn.hutool.core.util.CharUtil;
 import cn.hutool.core.util.ClassUtil;
 import cn.hutool.core.util.StrUtil;
 
@@ -79,6 +80,9 @@ public abstract class AbstractConverter<T> implements Converter<T> {
 			return (String) value;
 		} else if (ArrayUtil.isArray(value)) {
 			return ArrayUtil.toString(value);
+		} else if(CharUtil.isChar(value)) {
+			//对于ASCII字符使用缓存加速转换，减少空间创建
+			return CharUtil.toString((char)value);
 		}
 		return value.toString();
 	}