+++ hutool-core/src/main/java/cn/hutool/core/util/CharUtil.java
@@ -1,0 +1,229 @@
+package cn.hutool.core.util;
+
+import cn.hutool.core.text.ASCIIStrCache;
+
+/**
+ * 字符工具类<br>
+ * 部分工具来自于Apache Commons系列
+ * 
+ * @author looly
+ * @since 4.0.1
+ */
+public class CharUtil {
+
+	public static final char SPACE = ' ';
+	public static final char TAB = '	';
+	public static final char DOT = '.';
+	public static final char SLASH = '/';
+	public static final char BACKSLASH = '\\';
+	public static final char CR = '\r';
+	public static final char LF = '\n';
+	public static final char UNDERLINE = '_';
+	public static final char COMMA = ',';
+	public static final char DELIM_START = '{';
+	public static final char DELIM_END = '}';
+	public static final char BRACKET_START = '[';
+	public static final char BRACKET_END = ']';
+	public static final char COLON = ':';
+
+	/**
+	 * 是否为ASCII字符，ASCII字符位于0~127之间
+	 *
+	 * <pre>
+	 *   CharUtil.isAscii('a')  = true
+	 *   CharUtil.isAscii('A')  = true
+	 *   CharUtil.isAscii('3')  = true
+	 *   CharUtil.isAscii('-')  = true
+	 *   CharUtil.isAscii('\n') = true
+	 *   CharUtil.isAscii('&copy;') = false
+	 * </pre>
+	 * 
+	 * @param ch 被检查的字符处
+	 * @return true表示为ASCII字符，ASCII字符位于0~127之间
+	 */
+	public static boolean isAscii(char ch) {
+		return ch < 128;
+	}
+
+	/**
+	 * 是否为可见ASCII字符，可见字符位于32~126之间
+	 *
+	 * <pre>
+	 *   CharUtil.isAsciiPrintable('a')  = true
+	 *   CharUtil.isAsciiPrintable('A')  = true
+	 *   CharUtil.isAsciiPrintable('3')  = true
+	 *   CharUtil.isAsciiPrintable('-')  = true
+	 *   CharUtil.isAsciiPrintable('\n') = false
+	 *   CharUtil.isAsciiPrintable('&copy;') = false
+	 * </pre>
+	 * 
+	 * @param ch 被检查的字符处
+	 * @return true表示为ASCII可见字符，可见字符位于32~126之间
+	 */
+	public static boolean isAsciiPrintable(char ch) {
+		return ch >= 32 && ch < 127;
+	}
+
+	/**
+	 * 是否为ASCII控制符（不可见字符），控制符位于0~31和127
+	 *
+	 * <pre>
+	 *   CharUtil.isAsciiControl('a')  = false
+	 *   CharUtil.isAsciiControl('A')  = false
+	 *   CharUtil.isAsciiControl('3')  = false
+	 *   CharUtil.isAsciiControl('-')  = false
+	 *   CharUtil.isAsciiControl('\n') = true
+	 *   CharUtil.isAsciiControl('&copy;') = false
+	 * </pre>
+	 * 
+	 * @param ch 被检查的字符
+	 * @return true表示为控制符，控制符位于0~31和127
+	 */
+	public static boolean isAsciiControl(final char ch) {
+		return ch < 32 || ch == 127;
+	}
+
+	/**
+	 * 判断是否为字母（包括大写字母和小写字母）<br>
+	 * 字母包括A~Z和a~z
+	 *
+	 * <pre>
+	 *   CharUtil.isLetter('a')  = true
+	 *   CharUtil.isLetter('A')  = true
+	 *   CharUtil.isLetter('3')  = false
+	 *   CharUtil.isLetter('-')  = false
+	 *   CharUtil.isLetter('\n') = false
+	 *   CharUtil.isLetter('&copy;') = false
+	 * </pre>
+	 * 
+	 * @param ch 被检查的字符
+	 * @return true表示为字母（包括大写字母和小写字母）字母包括A~Z和a~z
+	 */
+	public static boolean isLetter(char ch) {
+		return isLetterUpper(ch) || isLetterLower(ch);
+	}
+
+	/**
+	 * <p>
+	 * 判断是否为大写字母，大写字母包括A~Z
+	 * </p>
+	 *
+	 * <pre>
+	 *   CharUtil.isLetterUpper('a')  = false
+	 *   CharUtil.isLetterUpper('A')  = true
+	 *   CharUtil.isLetterUpper('3')  = false
+	 *   CharUtil.isLetterUpper('-')  = false
+	 *   CharUtil.isLetterUpper('\n') = false
+	 *   CharUtil.isLetterUpper('&copy;') = false
+	 * </pre>
+	 * 
+	 * @param ch 被检查的字符
+	 * @return true表示为大写字母，大写字母包括A~Z
+	 */
+	public static boolean isLetterUpper(final char ch) {
+		return ch >= 'A' && ch <= 'Z';
+	}
+
+	/**
+	 * <p>
+	 * 检查字符是否为小写字母，小写字母指a~z
+	 * </p>
+	 *
+	 * <pre>
+	 *   CharUtil.isLetterLower('a')  = true
+	 *   CharUtil.isLetterLower('A')  = false
+	 *   CharUtil.isLetterLower('3')  = false
+	 *   CharUtil.isLetterLower('-')  = false
+	 *   CharUtil.isLetterLower('\n') = false
+	 *   CharUtil.isLetterLower('&copy;') = false
+	 * </pre>
+	 * 
+	 * @param ch 被检查的字符
+	 * @return true表示为小写字母，小写字母指a~z
+	 */
+	public static boolean isLetterLower(final char ch) {
+		return ch >= 'a' && ch <= 'z';
+	}
+
+	/**
+	 * <p>
+	 * 检查是否为数字字符，数字字符指0~9
+	 * </p>
+	 *
+	 * <pre>
+	 *   CharUtil.isNumber('a')  = false
+	 *   CharUtil.isNumber('A')  = false
+	 *   CharUtil.isNumber('3')  = true
+	 *   CharUtil.isNumber('-')  = false
+	 *   CharUtil.isNumber('\n') = false
+	 *   CharUtil.isNumber('&copy;') = false
+	 * </pre>
+	 * 
+	 * @param ch 被检查的字符
+	 * @return true表示为数字字符，数字字符指0~9
+	 */
+	public static boolean isNumber(final char ch) {
+		return ch >= '0' && ch <= '9';
+	}
+
+	/**
+	 * 是否为字符或数字，包括A~Z、a~z、0~9
+	 *
+	 * <pre>
+	 *   CharUtil.isLetterOrNumber('a')  = true
+	 *   CharUtil.isLetterOrNumber('A')  = true
+	 *   CharUtil.isLetterOrNumber('3')  = true
+	 *   CharUtil.isLetterOrNumber('-')  = false
+	 *   CharUtil.isLetterOrNumber('\n') = false
+	 *   CharUtil.isLetterOrNumber('&copy;') = false
+	 * </pre>
+	 * 
+	 * @param ch 被检查的字符
+	 * @return true表示为字符或数字，包括A~Z、a~z、0~9
+	 */
+	public static boolean isLetterOrNumber(final char ch) {
+		return isLetter(ch) || isNumber(ch);
+	}
+
+	/**
+	 * 字符转为字符串<br>
+	 * 如果为ASCII字符，使用缓存
+	 * 
+	 * @param c 字符
+	 * @return 字符串
+	 * @see ASCIIStrCache#toString(char)
+	 */
+	public static String toString(char c) {
+		return ASCIIStrCache.toString(c);
+	}
+
+	/**
+	 * 给定类名是否为字符类，字符类包括：
+	 * 
+	 * <pre>
+	 * Character.class
+	 * char.class
+	 * </pre>
+	 * 
+	 * @param clazz 被检查的类
+	 * @return true表示为字符类
+	 */
+	public static boolean isCharClass(Class<?> clazz) {
+		return clazz == Character.class || clazz == char.class;
+	}
+
+	/**
+	 * 给定类名是否为字符类，字符类包括：
+	 * 
+	 * <pre>
+	 * Character.class
+	 * char.class
+	 * </pre>
+	 * 
+	 * @param clazz 被检查的类
+	 * @return true表示为字符类
+	 */
+	public static boolean isChar(Object value) {
+		return value instanceof Character || value.getClass() == char.class;
+	}
+}